commit paginacion

+++ G01_P02/src/main/java/com/grupo01/spring/dao/GameDAO.java
@@ -2,6 +2,8 @@ package com.grupo01.spring.dao;
 
 import java.util.List;
 
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
@@ -18,7 +20,7 @@ import com.grupo01.spring.model.Game;
  */
 
 @Repository
-public interface GameDAO extends JpaRepository<Game, Integer> {
+public interface GameDAO extends JpaRepository<Game, Long> {
 	
 	/**
 	 * Buscador de juegos filtrando por genero
